Refactored code Refactored code
Rafactored Code
Refactored code
