+++ applications/data_parser.py
@@ -2,6 +2,8 @@ from gsmote import GeometricSMOTE
 from sklearn.model_selection import train_test_split
 import pandas as pd
 import numpy as np
+from imblearn.over_sampling import SMOTE
+
 class InputParser:
 
     @staticmethod
@@ -10,9 +12,10 @@ class InputParser:
         df = pd.read_csv(filename)
         X = np.asarray(df.iloc[:, :-1].values)
         y = np.asarray(df.iloc[:, -1].values)
-
         X_t, X_test, y_t, y_test = train_test_split(X, y, test_size=0.2, random_state=0)
-        X_train, y_train = gsmote.fit_resample(X_t,y_t)
+        # X_train, y_train = gsmote.fit_resample(X_t,y_t)
+        smt = SMOTE()
+        X_train, y_train = smt.fit_sample(X_t, y_t)
         classes = y_train.tolist()
         labels = y_train.tolist()
         input_database = {
