+++ gsmote/__init__.py
@@ -3,8 +3,8 @@ The :mod:`gsmote` provides the implementation of
 Geometric SMOTE algorithm.
 """
 
-from .geometric_smote import GeometricSMOTE
+from .eg_smote import EGSmote
 
 from ._version import __version__
 
-__all__ = ['GeometricSMOTE', '__version__']
+__all__ = ['EGSmote', '__version__']
