+++ gsmote/comparison_testing/Evaluator.py
@@ -15,5 +15,11 @@ def evaluate(classifier,Y_test,y_pred):
     g_mean = math.sqrt(tp*tn/((tp+fn)*(tn+fp)))
     AUC = (tp/(tp+fn)+tn/(tn+fp))/2
 
+    print(classifier, " finished executing")
+    print("\nClassifier: " + classifier)
+    print("f_score: " + str(f_score))
+    print("g_mean: " + str(g_mean))
+    print("AUC value: " + str(AUC) + "\n")
+
     print(classifier," finished executing.")
     return [classifier,f_score,g_mean,AUC]
